Fixed a bunch of printf-related warnings; repared BZWParser::cutWhiteSpace so it doesn't segfault for 0-length strings
git-svn-id: https://bzflag.svn.sourceforge.net/svnroot/bzflag/trunk@16741 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ bzworkbench/include/DataEntry.h
@@ -86,7 +86,7 @@ public:
 	// helper method:  ensures that a key occurs only once (useful for calling from update())
 	bool hasOnlyOne(vector<string> lines, const char* key) {
 		if( lines.size() != 1) {
-			printf("%s::update():  Error!  Defined \"%s\" %d times!\n", header.c_str(), key, lines.size());
+			printf("%s::update():  Error!  Defined \"%s\" %d times!\n", header.c_str(), key, (int)lines.size());
 			return false;	
 		}
 		return true;
